+++ NEWS.md
@@ -1,3 +1,8 @@
+# crime 0.2.0
+
+* Adds a function to that returns a data.frame from a regression model
+so it's easier to turn into tables for publication.
+
 # crime 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.